+++ changelog.md
@@ -1,6 +1,7 @@
 #### 0.0.1 (2020-12-21)
 
-- Optimize Slug Fields (#ffb12c3)
+- add filter `acfml/convert_wp_urls` (#679b71f)
+- Optimize Slug Fields (#3e44b84)
 - use user_trailingslashit for $paths (#104228c)
 - Make post type archive and rewrite slugs multilingual (#0634864)
 - support for rewrite front (#f9b7724)